add user service, repo, dto

+++ src/main/java/com/DataBaseProject/PCenter/dto/ProductDto.java
@@ -5,17 +5,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class ProductDto {
     private Integer id;
     private @NotNull String name;
-    private @NotNull double price;
+    private @NotNull BigDecimal price;
     private @NotNull String description;
     private @NotNull Integer categoryId;
 
-    public ProductDto(String name, double price, String description, Integer categoryId){
+    public ProductDto(String name, BigDecimal price, String description, Integer categoryId){
         this.name = name;
         this.price = price;
         this.description = description;
